docs/requirements.txt
sphinx-substitution-extensions
sphinx-tabs
sphinx-toolbox
sphinxcontrib-jquery
sdeper